deletion of duplicated codes for minimize and maximize window actions
