+++ jhotdraw-actions/src/main/java/org/jhotdraw/action/window/MaximizeWindowAction.java
@@ -8,12 +8,6 @@
 package org.jhotdraw.action.window;
 
 import java.awt.*;
-import java.awt.event.*;
-import javax.swing.*;
-import org.jhotdraw.action.AbstractViewAction;
-import org.jhotdraw.api.app.Application;
-import org.jhotdraw.api.app.View;
-import org.jhotdraw.util.*;
 
 /**
  * Maximizes the window of the active view.
@@ -21,33 +15,19 @@ import org.jhotdraw.util.*;
  * @author Werner Randelshofer
  * @version $Id$
  */
-public class MaximizeWindowAction extends AbstractViewAction {
+public class MaximizeWindowAction extends AbstractMinimizeMaximizeAction {
 
-    private static final long serialVersionUID = 1L;
     public static final String ID = "window.maximize";
 
     /**
      * Creates a new instance.
      */
     public MaximizeWindowAction(Application app, View view) {
-        super(app, view);
-        ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.action.Labels");
-        labels.configureAction(this, ID);
-    }
-
-    private JFrame getFrame() {
-        return (JFrame) SwingUtilities.getWindowAncestor(
-                getActiveView().getComponent()
-        );
+        super(app, view, ID);
     }
 
     @Override
     public void actionPerformed(ActionEvent evt) {
-        JFrame frame = getFrame();
-        if (frame != null) {
-            frame.setExtendedState(frame.getExtendedState() ^ Frame.MAXIMIZED_BOTH);
-        } else {
-            Toolkit.getDefaultToolkit().beep();
-        }
+        super.actionPerformed(evt,Frame.MAXIMIZED_BOTH);
     }
 }
