+++ jhotdraw-actions/src/main/java/org/jhotdraw/action/window/MinimizeWindowAction.java
@@ -8,12 +8,6 @@
 package org.jhotdraw.action.window;
 
 import java.awt.*;
-import java.awt.event.*;
-import javax.swing.*;
-import org.jhotdraw.action.AbstractViewAction;
-import org.jhotdraw.api.app.Application;
-import org.jhotdraw.api.app.View;
-import org.jhotdraw.util.*;
 
 /**
  * Minimizes the Frame of the current view.
@@ -21,33 +15,19 @@ import org.jhotdraw.util.*;
  * @author Werner Randelshofer
  * @version $Id$
  */
-public class MinimizeWindowAction extends AbstractViewAction {
+public class MinimizeWindowAction extends AbstractMinimizeMaximizeAction {
 
-    private static final long serialVersionUID = 1L;
     public static final String ID = "window.minimize";
 
     /**
      * Creates a new instance.
      */
     public MinimizeWindowAction(Application app, View view) {
-        super(app, view);
-        ResourceBundleUtil labels = ResourceBundleUtil.getBundle("org.jhotdraw.action.Labels");
-        labels.configureAction(this, ID);
-    }
-
-    private JFrame getFrame() {
-        return (JFrame) SwingUtilities.getWindowAncestor(
-                getActiveView().getComponent()
-        );
+        super(app, view, ID);
     }
 
     @Override
     public void actionPerformed(ActionEvent evt) {
-        JFrame frame = getFrame();
-        if (frame != null) {
-            frame.setExtendedState(frame.getExtendedState() ^ Frame.ICONIFIED);
-        } else {
-            Toolkit.getDefaultToolkit().beep();
-        }
+        super.actionPerformed(evt,Frame.ICONIFIED);
     }
 }
